Generalize StackId type
Previously, it was only used in the context of a buildpack.toml. StackIds are used in other contexts too and this PR generalizes them for other use-cases. In addtion, the type is now used in BuildContext and DetectContext instead of using a String.
