+++ libcnb/src/detect.rs
@@ -1,13 +1,14 @@
 use std::fmt::Debug;
 use std::path::PathBuf;
 
+use crate::data::stack_id::StackId;
 use crate::{data::build_plan::BuildPlan, data::buildpack::BuildpackToml, platform::Platform};
 
 /// Context for a buildpack's detect phase execution.
 pub struct DetectContext<P: Platform, BM> {
     pub app_dir: PathBuf,
     pub buildpack_dir: PathBuf,
-    pub stack_id: String,
+    pub stack_id: StackId,
     pub platform: P,
     pub buildpack_descriptor: BuildpackToml<BM>,
 }
